nlp/requirements.txt
absl-py==1.3.0
astroid==2.4.2
astunparse==1.6.3
beautifulsoup4==4.11.1
bpemb==0.3.4
cachetools==5.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
conllu==4.5.2
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.28
Deprecated==1.2.13
filelock==3.8.0
flair==0.11.3
flatbuffers==22.10.26
fonttools==4.38.0
ftfy==6.1.1
future==0.18.2
gast==0.4.0
gdown==4.4.0
gensim==3.8.3
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.10.1
hyperopt==0.2.7
idna==3.4
importlib-metadata==3.10.1
isort==5.6.4
Janome==0.4.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
lazy-object-proxy==1.4.3
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mccabe==0.6.1
more-itertools==9.0.0
mpld3==0.3
networkx==2.8.8
nltk==3.7
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pptree==3.1
progressbar==2.5
protobuf==3.19.6
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
segtok==1.5.11
sentencepiece==0.1.95
six==1.16.0
smart-open==6.2.0
soupsieve==2.3.2.post1
sqlitedict==2.0.0
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
torch==1.13.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
Wikipedia-API==0.5.4
wincertstore==0.2
wrapt==1.12.1
zipp==3.10.0